@@ -13,23 +13,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0-rc.7",
-    "@angular/common": "^5.0.0-rc.7",
-    "@angular/compiler": "^5.0.0-rc.7",
-    "@angular/core": "^5.0.0-rc.7",
-    "@angular/forms": "^5.0.0-rc.7",
-    "@angular/http": "^5.0.0-rc.7",
-    "@angular/platform-browser": "^5.0.0-rc.7",
-    "@angular/platform-browser-dynamic": "^5.0.0-rc.7",
-    "@angular/router": "^5.0.0-rc.7",
+    "@angular/animations": "^6.1.10",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
+    "@angular/platform-browser": "^6.1.10",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/router": "^6.1.10",
     "core-js": "^2.4.1",
-    "rxjs": "^5.5.0",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.6.2",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0-rc.5",
-    "@angular/compiler-cli": "^5.0.0-rc.7",
-    "@angular/language-service": "^5.0.0-rc.7",
-    "typescript": "2.5.3"
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "6",
+    "@angular/compiler-cli": "^6.1.10",
+    "@angular/language-service": "^6.1.10",
+    "typescript": "2.9.2"
   }
 }
