@@ -1,18 +1,10 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
+import {map, shareReplay, merge, mergeMap, scan} from 'rxjs/operators';
+import {Injectable} from '@angular/core';
+import {HttpClient} from '@angular/common/http';
 
-import { RootObject, Person, Ship, Movie } from '../interfaces';
+import {RootObject, Person, Ship, Movie} from '../interfaces';
 
-import { of } from 'rxjs/observable/of';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/merge';
-import 'rxjs/add/operator/mergeMap';
-import 'rxjs/add/operator/toPromise';
-import 'rxjs/add/operator/scan';
-import 'rxjs/add/operator/shareReplay';
-import 'rxjs/add/operator/switchMap';
+import {of, Observable} from 'rxjs';
 
 @Injectable()
 export class SwapiService {
@@ -21,35 +13,30 @@ export class SwapiService {
   private _load = <T>(url) =>
     this.http
       .get<RootObject<T>>(url)
-      .mergeMap(
-        root => (root.next ? of(root).merge(this._load(root.next)) : of(root))
-      );
+      .pipe(mergeMap((root) => (root.next ? of(root).pipe(merge(this._load(root.next.replace('http:', 'https:')))) : of(root))));
 
   public loadAll = <T>(url): Observable<T[]> =>
-    this._load<T>(url)
-      .map((r: RootObject<T>) => r.results)
-      .scan((combinedList, latestAdditions) =>
-        combinedList.concat(latestAdditions)
-      )
-      .map(set => set.sort((x, y) => (x.name < y.name ? -1 : 1)));
+    this._load<T>(url).pipe(
+      map((r: RootObject<T>) => r.results),
+      scan((combinedList:T[], latestAdditions) => combinedList.concat(latestAdditions)),
+      map((set:any[]) => set.sort((x, y) => (x.name < y.name ? -1 : 1)))
+    );
 }
 
 @Injectable()
 export class PeopleService {
   constructor(private swapi: SwapiService) {}
 
-  persons$ = this.swapi.loadAll<Person>('https://swapi.co/api/people/');
+  persons$ = this.swapi.loadAll<Person>('https://swapi.dev/api/people/');
 }
 
 @Injectable()
 export class StarshipsService {
-  private starships$ = this.swapi
-    .loadAll<Ship>('https://swapi.co/api/starships')
-    .shareReplay(1);
+  private starships$ = this.swapi.loadAll<Ship>('https://swapi.dev/api/starships').pipe(shareReplay(1));
 
   load(url) {
     // this version doesn't load the ships over and over again anymore
-    return this.starships$.map(ships => ships.find(cur => cur.url === url));
+    return this.starships$.pipe(map((ships) => ships.find((cur) => cur.url === url)));
   }
   constructor(private swapi: SwapiService) {
     // trick to load the straships at init, and keep them in mmory
@@ -61,13 +48,11 @@ export class StarshipsService {
 
 @Injectable()
 export class MovieService {
-  private movies$ = this.swapi
-    .loadAll<Movie>('https://swapi.co/api/films')
-    .shareReplay(1);
+  private movies$ = this.swapi.loadAll<Movie>('https://swapi.dev/api/films').pipe(shareReplay(1));
 
   load(url) {
     // this version doesn't load the ships over and over again anymore
-    return this.movies$.map(ships => ships.find(cur => cur.url === url));
+    return this.movies$.pipe(map((ships) => ships.find((cur) => cur.url === url)));
   }
   constructor(private swapi: SwapiService) {
     // trick to load the straships at init, and keep them in mmory
