@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Movie } from '../../../interfaces';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-movies',
